Fix pathing after refactoring into versions
